fix sock_cb to add free file
