+++ kernel/fs/file.c
@@ -67,12 +67,8 @@ filedup(struct file *f)
   return f;
 }
 
-// Close file f.  (Decrement ref count, close when reaches 0.)
-void
-fileclose(struct file *f)
+void filefree(struct file *f)
 {
-  struct file ff;
-
   acquire(&ftable.lock);
   if(f->ref < 1)
     panic("fileclose");
@@ -80,7 +76,6 @@ fileclose(struct file *f)
     release(&ftable.lock);
     return;
   }
-  ff = *f;
   f->ref = 0;
   // f->type = FD_NONE;
 
@@ -93,8 +88,15 @@ fileclose(struct file *f)
     f->next->prev = f->prev;
   }
   bd_free(f);
-
   release(&ftable.lock);
+}
+
+// Close file f.  (Decrement ref count, close when reaches 0.)
+void
+fileclose(struct file *f)
+{
+  struct file ff;
+  ff = *f;
 
   if(ff.type == FD_PIPE){
     pipeclose(ff.pipe, ff.writable);
@@ -103,6 +105,7 @@ fileclose(struct file *f)
     iput(ff.ip);
     end_op();
   } else if (ff.type == FD_SOCK) {
+    ff.scb->f = 0;
     sockclose(&ff);
   }
 }
