+++ kernel/include/file.h
@@ -47,3 +47,12 @@ struct devsw {
 extern struct devsw devsw[];
 
 #define CONSOLE 1
+
+struct file*    filealloc(void);
+void            filefree(struct file*);
+void            fileclose(struct file*);
+struct file*    filedup(struct file*);
+void            fileinit(void);
+int             fileread(struct file*, uint64, int n);
+int             filestat(struct file*, uint64 addr);
+int             filewrite(struct file*, uint64, int n);
